requirements.txt
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv
requests
beautifulsoup4
PyPDF2
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.0
numpy
scikit-learn
torch==2.2.2
psycopg2-binary
langchain
langchain-community
unstructured
gunicorn
dj-database-url
celery
redis
gevent==21.12.0

